Sprites change

+++ include/libft.h
@@ -6,7 +6,7 @@
 /*   By: lmartins <lmartins@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/30 11:51:35 by lmartins          #+#    #+#             */
-/*   Updated: 2021/06/20 07:02:22 by lmartins         ###   ########.fr       */
+/*   Updated: 2021/06/27 08:31:44 by lmartins         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -83,7 +83,7 @@ size_t				ft_intlen(int num);
 size_t				ft_hexlen(size_t nbr);
 char				*ft_itoa_uns(unsigned int num);
 char				*ft_itoa_base(size_t nb, int base);
-void				ft_swap(float *a, float *b);
+void				ft_swap(double *a, double *b);
 void				ft_replace(char *str, char find, char replace);
 char				*ft_strjoin_free(char *s1, char *s2);
 void				*ft_calloc_char(size_t count, char c);
